#103 - Merge Smaller Generators into a Skeleton Generator
Fixes #103
Fixes #104
Fixes #105
