--- generators/client-assets/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../../util/common-generator').generator(__dirname);
