--- generators/client-build/phases/writing.js
@@ -1,27 +1,0 @@
-module.exports = {
-  packageJson() {
-    this.addDevDependencies(
-      'babelify',
-      'mkdirp',
-      'browserify',
-      'browserify-shim',
-      'envify',
-      'uglifyify',
-      'watchify',
-      'jsdom',
-      'gulp-livereload',
-      'gulp-rev-all',
-      'gulp-awspublish',
-      'gulp-cloudfront',
-      'gulp-sourcemaps',
-      'gulp-cssmin',
-      'gulp-imagemin',
-      'gulp-jade'
-    );
-
-    // React-Specific build dependencies
-    if (this.props.client === 'react') {
-      this.addDevDependency('livereactload');
-    }
-  },
-};
