--- generators/client-styles/phases/writing.js
@@ -1,5 +1,0 @@
-module.exports = {
-  packageJson() {
-    this.addDevDependency('gulp-sass');
-  },
-};
