--- generators/eslint/phases/writing.js
@@ -1,10 +1,0 @@
-module.exports = {
-  packageJson() {
-    this.addDevDependencies(
-      'gulp-eslint',
-      'babel-eslint',
-      'eslint-config-airbnb',
-      'eslint-plugin-react'
-    );
-  },
-};
