--- generators/git-hooks/phases/writing.js
@@ -1,6 +1,0 @@
-module.exports = {
-  packageJson() {
-    this.updateNpmScript('postinstall', 'npm run install-git-hooks');
-    this.updateNpmScript('install-git-hooks', './scripts/npm/install-git-hooks.sh');
-  },
-};
