--- generators/gulp/templates/gulp/tasks/default.js
@@ -1,12 +1,0 @@
-const gulp = require('gulp');
-const runSequence = require('run-sequence');
-
-gulp.task('default', (cb) => {
-  return runSequence(
-    'lint',
-    'prepare-assets',
-    'bundle-client',
-    'test',
-    cb
-  );
-});
