--- generators/server-common/index.js
@@ -1,3 +1,0 @@
-module.exports = require('../../util/common-generator').generator(__dirname, {
-  writing: require('./phases/writing'),
-});
