--- generators/server-common/phases/writing.js
@@ -1,20 +1,0 @@
-module.exports = {
-  packageJson() {
-    this.addDependencies(
-      'babel',
-      'bcryptjs',
-      'bluebird',
-      'coffee-script',
-      'config',
-      'debug',
-      'jade',
-      'express',
-      'lodash'
-    );
-
-    this.addDevDependencies(
-      'chai',
-      'gulp-nodemon'
-    );
-  },
-};
