+++ generators/skeleton/index.js
