+++ generators/skeleton/templates/Dockerfile
