+++ generators/skeleton/templates/README.md.tpl
