+++ generators/skeleton/templates/client/assets/favicon.ico
