+++ generators/skeleton/templates/client/styles/_reset.scss
