+++ generators/skeleton/templates/client/styles/_variables.scss
