+++ generators/skeleton/templates/client/styles/main.scss
