+++ generators/skeleton/templates/docker-compose.yml
