+++ generators/skeleton/templates/editorconfig
