+++ generators/skeleton/templates/eslintrc
