+++ generators/skeleton/templates/gitattributes
