+++ generators/skeleton/templates/gitignore
