+++ generators/skeleton/templates/gulp/config.js.tpl
@@ -1,0 +1,39 @@
+const _ = require('lodash');
+
+function sourceNode(rootName, root, extra = {}) {
+  return _.merge({
+    all: [root + '/**/*.js'],
+    source: [root + '/**/*.js', '!' + root + '/**/*.spec.js', '!' + root + '/**/*spec*/*'],
+    test: [root + '/**/*.spec.js', root + '/**/*spec*/*'],
+    output: {
+      coverage: 'target/test-reports/' + rootName,
+    },
+  }, extra);
+}
+
+module.exports = {
+  server: sourceNode('server', 'server'),
+  all: sourceNode('all', '{client,server}'),
+  build: ['gulpfile.js', 'gulp/**/*.js'],
+  client: sourceNode('client', 'client', {
+    styles: ['client/**/*.{scss,sass}'],
+    images: ['client/**/*.{gif,jpeg,jpg,png,svg}'],
+    staticJade: ['client/**/*.jade', '!client/**/*.dynamic.jade'],
+    assets: ['client/assets/**/*.*'],
+    html: ['client/**/*.html'],
+    entries: ['client/app.js'],
+    dist: {
+      path: 'public',
+      styles: 'public/styles',
+      assets: 'public/assets',
+      bundle: 'app.js',
+    },
+  }),
+  imagemin: { optimizationLevel: 4 },
+<% if (client === 'angular') { %>
+  ngHtml2Js: {
+    module: 'templates',
+    baseDir: 'public',
+  },
+<% } %>
+};
