+++ generators/skeleton/templates/gulp/index.js
