+++ generators/skeleton/templates/gulp/tasks/server.js
