+++ generators/skeleton/templates/gulp/tasks/test.js
