+++ generators/skeleton/templates/gulpfile.js
