+++ generators/skeleton/templates/jsconfig.json
