+++ generators/skeleton/templates/nodemonignore
