+++ generators/skeleton/templates/scripts/git/hooks/commit-msg
