+++ generators/skeleton/templates/scripts/npm/install-git-hooks.sh
